+++ frontend-plugin-core/src/main/java/com/github/eirslett/maven/plugins/frontend/lib/NpmRunner.java
@@ -18,7 +18,8 @@ final class DefaultNpmRunner extends NodeTaskExecutor implements NpmRunner {
         		buildProxy(proxyConfig, npmRegistryURL));
     }
 
-    private static List<String> buildArguments(ProxyConfig proxyConfig, String npmRegistryURL) {
+    // Visible for testing only.
+    static List<String> buildArguments(ProxyConfig proxyConfig, String npmRegistryURL) {
         List<String> arguments = new ArrayList<String>();
 
         if(npmRegistryURL != null && !npmRegistryURL.isEmpty()){
@@ -30,6 +31,14 @@ final class DefaultNpmRunner extends NodeTaskExecutor implements NpmRunner {
 
             arguments.add("--https-proxy=" + proxy.getUri().toString());
             arguments.add("--proxy=" + proxy.getUri().toString());
+
+            final String nonProxyHosts = proxy.getNonProxyHosts();
+            if (nonProxyHosts != null && !nonProxyHosts.isEmpty()) {
+                final String[] nonProxyHostList = nonProxyHosts.split("\\|");
+                for (String nonProxyHost: nonProxyHostList) {
+                    arguments.add("--noproxy=" + nonProxyHost.replace("*", ""));
+                }
+            }
         }
 
         return arguments;