+++ frontend-plugin-core/src/main/java/com/github/eirslett/maven/plugins/frontend/lib/YarnInstaller.java
@@ -1,5 +1,6 @@
 package com.github.eirslett.maven.plugins.frontend.lib;
 
+import java.io.EOFException;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -22,6 +23,8 @@ public class YarnInstaller {
 
     private String yarnVersion, yarnDownloadRoot, userName, password;
 
+    private boolean isYarnBerry;
+
     private final Logger logger;
 
     private final InstallConfig config;
@@ -39,6 +42,11 @@ public class YarnInstaller {
 
     public YarnInstaller setYarnVersion(String yarnVersion) {
         this.yarnVersion = yarnVersion;
+        return this;
+    }
+
+    public YarnInstaller setIsYarnBerry(boolean isYarnBerry) {
+        this.isYarnBerry = isYarnBerry;
         return this;
     }
 
@@ -84,8 +92,13 @@ public class YarnInstaller {
                     logger.info("Yarn {} is already installed.", version);
                     return true;
                 } else {
-                    logger.info("Yarn {} was installed, but we need version {}", version, yarnVersion);
-                    return false;
+                    if (isYarnBerry && Integer.parseInt(version.split("\\.")[0]) > 1) {
+                        logger.info("Yarn Berry {} is installed.", version);
+                        return true;
+                    } else{
+                        logger.info("Yarn {} was installed, but we need version {}", version, yarnVersion);
+                        return false;
+                    }
                 }
             } else {
                 return false;
@@ -122,7 +135,23 @@ public class YarnInstaller {
                 logger.warn("Failed to delete existing Yarn installation.");
             }
 
-            extractFile(archive, installDirectory);
+            try {
+                extractFile(archive, installDirectory);
+            } catch (ArchiveExtractionException e) {
+                if (e.getCause() instanceof EOFException) {
+                    // https://github.com/eirslett/frontend-maven-plugin/issues/794
+                    // The downloading was probably interrupted and archive file is incomplete:
+                    // delete it to retry from scratch
+                    this.logger.error("The archive file {} is corrupted and will be deleted. "
+                            + "Please try the build again.", archive.getPath());
+                    archive.delete();
+                    if (installDirectory.exists()) {
+                        FileUtils.deleteDirectory(installDirectory);
+                    }
+                }
+
+                throw e;
+            }
 
             ensureCorrectYarnRootDirectory(installDirectory, yarnVersion);